Override the baseURL when previewing
Currently, when running 'make preview' the generated pages contain links
to the deployed site, so it's not possible to actually preview local
changes.

Set the --baseURL when running 'hugo server' to specify the localhost
baseURL when running 'make preview'.

Add the clean target as a dependency to the 'preview' and 'deploy'
targets to be sure we don't preview or deploy the wrong baseURL.